engine return value - expand from 4-bit to 8-bit.
The original PA-RISC code used the deposit immediate instruction which was limited to a 4-bit constant (5-bits really, but one was the sign bit).  For toolbox break support, we need another return value so 4 bits isn't enough.

(In truth, the 3 RET_C7xx codes can be merged but that's a separate PR and the MVP/MVN aren't used by the C code)

the PA-RISC changes haven't been tested and I suspect that code hasn't been compiled this century but they look reasonable to me.